functions/func-simiriki-leadscore/requirements.txt
azure-functions==1.20.0
azure-storage-blob==12.20.0
